[ancillary.get_geometry] outsourced from ID.geometry; added and modified tests
