+++ tests/test_drivers.py
@@ -1,5 +1,4 @@
 import pyroSAR
-from pyroSAR.ancillary import get_corners
 
 import pytest
 import platform
@@ -26,7 +25,6 @@ def testcases():
             'bbox_area': 7.573045244595988,
             'compression': 'zip',
             'corners': {'ymax': 52.183979, 'ymin': 50.295261, 'xmin': 8.017178, 'xmax': 12.0268},
-            'hgt_len': 15,
             'lines': 16685,
             'orbit': 'A',
             'outname': 'S1A__IW___A_20150222T170750',
@@ -42,7 +40,6 @@ def testcases():
             'acquisition_mode': 'FBD',
             'compression': 'zip',
             'corners': {'xmin': -62.9005207, 'xmax': -62.1629744, 'ymin': -11.4233051, 'ymax': -10.6783401},
-            'hgt_len': 2,
             'lines': 13160,
             'orbit': 'A',
             'outname': 'PSR2_FBD__A_20140909T043342',
@@ -70,7 +67,7 @@ class Test_Metadata():
     def test_attributes(self, scene):
         assert scene['pyro'].acquisition_mode == scene['acquisition_mode']
         assert scene['pyro'].compression == scene['compression']
-        assert scene['pyro'].getCorners() == scene['corners']
+        assert scene['pyro'].getCorners() == pytest.approx(scene['corners'])
         assert scene['pyro'].lines == scene['lines']
         assert scene['pyro'].outname_base() == scene['outname']
         assert scene['pyro'].orbit == scene['orbit']
@@ -81,7 +78,6 @@ class Test_Metadata():
         assert scene['pyro'].stop == scene['stop']
         assert scene['pyro'].sensor == scene['sensor']
         assert scene['pyro'].spacing == scene['spacing']
-        assert len(scene['pyro'].getHGT()) == scene['hgt_len']
 
 
 def test_identify_fail(testdir, testdata):
@@ -110,53 +106,6 @@ def test_parse_date():
 
 def test_export2dict():
     pass
-
-
-@pytest.mark.parametrize(
-    'coordinates, expected',
-    [
-        (
-                [
-                    (10, 50),
-                    (10, 51),
-                    (11, 51),
-                    (11, 50),
-                ],
-                {'xmin': 10, 'xmax': 11, 'ymin': 50, 'ymax': 51},
-        ),
-        (
-                [
-                    (179, 50),
-                    (179, 51),
-                    (-179, 51),
-                    (-179, 50),
-                ],
-                {'xmin': 179, 'xmax': -179, 'ymin': 50, 'ymax': 51},
-        ),
-        (
-                [
-                    (175, 52),
-                    (-178, 50),
-                    (179, 51),
-                    (-170, 53),
-                    (178, 49),
-                ],
-                {'xmin': 175, 'xmax': -170, 'ymin': 49, 'ymax': 53},
-        ),
-        (
-                [
-                    (11, 51),
-                    (10, 50),
-                    (12, 52),
-                    (10.5, 50.5),
-                ],
-                {'xmin': 10, 'xmax': 12, 'ymin': 50, 'ymax': 52},
-        ),
-    ],
-    ids=['ordinary', 'antimeridian', 'unordered-antimeridian', 'extra-points'],
-)
-def test_get_corners(coordinates, expected):
-    assert get_corners(coordinates) == expected
 
 
 def test_getFileObj(tmpdir, testdata):
@@ -199,8 +148,10 @@ def test_scene(tmpdir, testdata):
         id.getGammaImages()
     assert id.getGammaImages(id.scene) == []
     id = pyroSAR.identify(testdata['psr2'])
-    assert id.getCorners() == {'xmax': -62.1629744, 'xmin': -62.9005207,
-                               'ymax': -10.6783401, 'ymin': -11.4233051}
+    assert id.getCorners() == pytest.approx({
+        'xmax': -62.1629744, 'xmin': -62.9005207,
+        'ymax': -10.6783401, 'ymin': -11.4233051
+    })
 
 
 datasets = ['asar', 'ers1_esa', 'ers1_ceos', 'psr2', 's1']
